pyproject.toml
[tool.poetry]
name = "smog"
version = "0.1.0"
description = "Airtable employee lookup client"
authors = ["Chase Pettet <chase@owasp.org>"]
readme = "README.md"
packages = [{include = "smog", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
pyairtable = "^2.3.3"
click = "^8.1.7"
pyyaml = "^6.0.1"
pydantic = "^2.6.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
mypy = "^1.8.0"
types-pyyaml = "^6.0.12"

[tool.poetry.scripts]
smog-lookup = "smog.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = true
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
